perf(decode): narrow mi-size grid element type from usize to u8
Feature ID: INFRA-DECODE-SERIAL-HOT-PATHS

Allocation campaign slice 3 (traffic). The coalesced TileMiSizeState backing
store holds block-size indices (< 29) and partition-context values (<= 63),
which fit in a byte but were stored as usize -- 8x wider than needed. Narrow
the storage to Vec<u8>. The partition-decision read path
(TilePartitionContextState, PartitionContextInput, SquareSplitContextInput,
and the grid/neighbour leaf helpers in cdf/context) now carries &[u8] and
widens to usize at the point of use; the block-size index is converted with a
checked u8::try_from that surfaces a typed error, preserving the no-panic
discipline. Values round-trip identically, so decode output is unchanged.

Measured (single-thread, in-container; no dav2d/1080p test.ivf available):
- DHAT allocation traffic: TIP 5,777,818 -> 5,697,178 bytes (-80.6 KB,
  -1.4%); opfl 3,682,490 -> 3,650,234 (-32.3 KB, -0.88%). Block count
  unchanged (same allocations, narrower). The saving scales with tile mi
  area, so it is larger at real resolutions than on these micro-fixtures.
- callgrind Ir: TIP -0.04%, opfl -0.05% (u8 loads/stores are cheaper).
- Byte-exact on all five fixtures; splot-decode suite green (1,643 tests).

+++ crates/splot-decode/src/bitstream/tile_payload/partition_traversal.rs
@@ -64,10 +64,10 @@ pub(crate) const BLOCK_32X8: usize = 22;
 #[allow(clippy::struct_field_names)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub(crate) struct TilePartitionContextState<'a> {
-    mi_sizes: &'a [usize],
+    mi_sizes: &'a [u8],
     mi_size_stride: usize,
-    left_mi_sizes: [&'a [usize]; 2],
-    above_mi_sizes: [&'a [usize]; 2],
+    left_mi_sizes: [&'a [u8]; 2],
+    above_mi_sizes: [&'a [u8]; 2],
     origin_row: usize,
     origin_col: usize,
 }
@@ -75,10 +75,10 @@ pub(crate) struct TilePartitionContextState<'a> {
 impl<'a> TilePartitionContextState<'a> {
     #[must_use]
     pub(crate) const fn new(
-        mi_sizes: &'a [usize],
+        mi_sizes: &'a [u8],
         mi_size_stride: usize,
-        left_mi_sizes: [&'a [usize]; 2],
-        above_mi_sizes: [&'a [usize]; 2],
+        left_mi_sizes: [&'a [u8]; 2],
+        above_mi_sizes: [&'a [u8]; 2],
     ) -> Self {
         Self::new_at(
             mi_sizes,
@@ -92,10 +92,10 @@ impl<'a> TilePartitionContextState<'a> {
 
     #[must_use]
     pub(crate) const fn new_at(
-        mi_sizes: &'a [usize],
+        mi_sizes: &'a [u8],
         mi_size_stride: usize,
-        left_mi_sizes: [&'a [usize]; 2],
-        above_mi_sizes: [&'a [usize]; 2],
+        left_mi_sizes: [&'a [u8]; 2],
+        above_mi_sizes: [&'a [u8]; 2],
         origin_row: usize,
         origin_col: usize,
     ) -> Self {
